packages/kaori-flow/pyproject.toml
[project]
name = "kaori-flow"
version = "2.0.0"
description = "Kaori Protocol — Trust Physics & Orchestration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "MSRO", email = "protocol@kaori.dev"}
]

dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kaori_flow"]
